requirements.txt
attrs==23.2.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
gspread==6.1.0
h11==0.14.0
idna==3.6
numpy==1.26.4
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.3.0.post0
pandas==2.2.1
patool==2.2.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
selenium==4.19.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
StrEnum==0.4.15
trio==0.25.0
trio-websocket==0.11.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vboxapi==1.0
wsproto==1.2.0
XlsxWriter==3.2.0
